[qol/fix] - flatten icon imports - improved icon access, simplified fs structure - fixed config files (package.json and tsconfig) - fixed circular reference for themes object and icons - correct name for XMark component
