+++ package.json
@@ -27,10 +27,10 @@
         "lucide-react": "^0.288.0",
         "mongodb": "^6.2.0",
         "mongoose": "^8.0.0",
-        "next": "latest",
+        "next": "^14.0.3",
         "next-auth": "^4.24.4",
-        "react": "latest",
-        "react-dom": "latest",
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
         "tailwind-merge": "^1.14.0",
         "tailwindcss-animate": "^1.0.7",
         "tone": "^14.7.77"
@@ -39,15 +39,15 @@
         "@formatjs/intl-localematcher": "^0.4.2",
         "@types/bcrypt": "^5.0.1",
         "@types/negotiator": "^0.6.1",
-        "@types/node": "latest",
-        "@types/react": "latest",
-        "@types/react-dom": "latest",
+        "@types/node": "^20.10.1",
+        "@types/react": "^18.2.39",
+        "@types/react-dom": "^18.2.17",
         "@typescript-eslint/eslint-plugin": "^6.11.0",
         "@typescript-eslint/parser": "^6.11.0",
         "autoprefixer": "^10.4.16",
         "cypress": "^13.6.0",
-        "eslint": "latest",
-        "eslint-config-next": "latest",
+        "eslint": "^8.54.0",
+        "eslint-config-next": "^14.0.3",
         "eslint-plugin-react": "^7.33.2",
         "negotiator": "^0.6.3",
         "postcss": "^8.4.31",
