+++ src/components/icons/PortfolioLogo.tsx
@@ -1,0 +1,25 @@
+import { ICON_COLORS } from '@/constants';
+import { IconProps } from '@/types';
+
+export const PortfolioLogo = ({
+    color = 'primary',
+    size = 'medium',
+}: IconProps) => {
+    const sizes = {
+        small: 'w-[112px] h-[18px]',
+        medium: 'w-[130px] h-[25px]',
+        large: 'w-[178px] h-[34px]',
+    };
+    return (
+        <svg
+            viewBox='0 0 178 34'
+            fill='none'
+            xmlns='http://www.w3.org/2000/svg'
+            className={`${ICON_COLORS[color]}  ${sizes[size]}`}>
+            <path
+                d='M11.44 5.496C14.032 5.496 16.224 5.736 18.016 6.216C19.808 6.696 21.232 7.624 22.288 9C23.376 10.344 23.92 12.28 23.92 14.808C23.92 15.928 23.744 17.08 23.392 18.264C23.072 19.416 22.656 20.424 22.144 21.288C21.664 22.152 21.232 22.712 20.848 22.968C20.4 23.288 19.408 23.544 17.872 23.736C16.336 23.896 14.528 23.976 12.448 23.976C11.712 23.976 10.576 23.944 9.04 23.88V33H0.4V5.64C3.76 5.544 7.44 5.496 11.44 5.496ZM15.04 17.832C15.36 17.416 15.52 16.584 15.52 15.336C15.52 13.704 15.168 12.552 14.464 11.88C13.792 11.208 12.608 10.872 10.912 10.872C10.336 10.872 9.712 10.92 9.04 11.016V18.312C9.36 18.344 9.984 18.36 10.912 18.36C13.376 18.36 14.752 18.184 15.04 17.832ZM37.2021 33.624C33.3621 33.624 30.4821 32.728 28.5621 30.936C26.6741 29.112 25.7301 26.248 25.7301 22.344C25.7301 20.2 25.9221 18.328 26.3061 16.728C26.7221 15.096 27.2181 13.96 27.7941 13.32C28.5301 12.488 29.8741 11.8 31.8261 11.256C33.8101 10.712 35.6661 10.44 37.3941 10.44C41.5221 10.44 44.4661 11.32 46.2261 13.08C47.9861 14.84 48.8661 17.656 48.8661 21.528C48.8661 25.688 47.9221 28.744 46.0341 30.696C44.1461 32.648 41.2021 33.624 37.2021 33.624ZM37.3461 28.008C38.5941 28.008 39.4101 27.576 39.7941 26.712C40.1781 25.816 40.3701 24.248 40.3701 22.008C40.3701 20.44 40.2901 19.24 40.1301 18.408C39.9701 17.544 39.6661 16.92 39.2181 16.536C38.8021 16.12 38.1461 15.912 37.2501 15.912C35.9701 15.912 35.2341 16.024 35.0421 16.248C34.7861 16.472 34.5781 17.176 34.4181 18.36C34.2581 19.544 34.1781 20.776 34.1781 22.056C34.1781 24.232 34.3861 25.768 34.8021 26.664C35.2501 27.56 36.0981 28.008 37.3461 28.008ZM65.6793 10.44C66.0953 10.44 66.5273 10.488 66.9753 10.584C67.4553 10.68 67.8073 10.792 68.0313 10.92C68.0313 11.624 67.9513 12.792 67.7913 14.424C67.6313 16.024 67.4873 17.16 67.3593 17.832C67.0393 17.8 66.4473 17.784 65.5833 17.784C63.5673 17.784 61.7753 17.896 60.2073 18.12V33H51.9993V10.92H58.8633L59.8233 14.472H59.8713C60.7993 13.256 61.7913 12.28 62.8473 11.544C63.9033 10.808 64.8473 10.44 65.6793 10.44ZM80.3628 23.928C80.3628 24.504 80.3788 25.112 80.4108 25.752C80.4428 26.36 80.5388 26.744 80.6988 26.904C80.9228 27.128 81.7068 27.24 83.0508 27.24L84.6828 27.192C84.9708 28.728 85.1148 30.504 85.1148 32.52C84.5068 32.744 83.6108 32.952 82.4268 33.144C81.2748 33.336 80.1708 33.432 79.1148 33.432C76.9068 33.432 75.1948 32.952 73.9788 31.992C72.7948 31 72.2028 29.48 72.2028 27.432V15.624H68.6028V11.256L72.2028 10.92L74.3628 5.64L80.3628 5.304V10.92H85.4508V15.624H80.3628V23.928ZM98.3695 8.424C98.1455 8.648 98.0335 9.336 98.0335 10.488V10.92H103.122V15.624H98.0335V33H89.8255V15.624H86.2255V11.352L89.8255 10.92V10.2C89.8255 8.792 90.0175 7.448 90.4015 6.168C90.7855 4.888 91.1375 4.12 91.4575 3.864C92.4495 3.032 94.5615 2.616 97.7935 2.616C99.5535 2.616 101.522 2.744 103.698 3C103.698 3.384 103.634 4.152 103.506 5.304C103.41 6.456 103.314 7.432 103.218 8.232L101.154 8.184C99.4575 8.184 98.5295 8.264 98.3695 8.424ZM115.858 33.624C112.018 33.624 109.138 32.728 107.218 30.936C105.33 29.112 104.386 26.248 104.386 22.344C104.386 20.2 104.578 18.328 104.962 16.728C105.378 15.096 105.874 13.96 106.45 13.32C107.186 12.488 108.53 11.8 110.482 11.256C112.466 10.712 114.322 10.44 116.05 10.44C120.178 10.44 123.122 11.32 124.882 13.08C126.642 14.84 127.522 17.656 127.522 21.528C127.522 25.688 126.578 28.744 124.69 30.696C122.802 32.648 119.858 33.624 115.858 33.624ZM116.002 28.008C117.25 28.008 118.066 27.576 118.45 26.712C118.834 25.816 119.026 24.248 119.026 22.008C119.026 20.44 118.946 19.24 118.786 18.408C118.626 17.544 118.322 16.92 117.874 16.536C117.458 16.12 116.802 15.912 115.906 15.912C114.626 15.912 113.89 16.024 113.698 16.248C113.442 16.472 113.234 17.176 113.074 18.36C112.914 19.544 112.834 20.776 112.834 22.056C112.834 24.232 113.042 25.768 113.458 26.664C113.906 27.56 114.754 28.008 116.002 28.008ZM130.8 4.2L139.008 3.864V33H130.8V4.2ZM147.349 8.184C146.261 8.184 145.317 7.816 144.517 7.08C143.749 6.344 143.365 5.464 143.365 4.44C143.365 3.416 143.749 2.536 144.517 1.8C145.317 1.064 146.261 0.695998 147.349 0.695998C148.437 0.695998 149.365 1.064 150.133 1.8C150.901 2.536 151.285 3.416 151.285 4.44C151.285 5.464 150.901 6.344 150.133 7.08C149.365 7.816 148.437 8.184 147.349 8.184ZM143.221 10.92H151.429V33H143.221V10.92ZM166.155 33.624C162.315 33.624 159.435 32.728 157.515 30.936C155.627 29.112 154.683 26.248 154.683 22.344C154.683 20.2 154.875 18.328 155.259 16.728C155.675 15.096 156.171 13.96 156.747 13.32C157.483 12.488 158.827 11.8 160.779 11.256C162.763 10.712 164.619 10.44 166.347 10.44C170.475 10.44 173.419 11.32 175.179 13.08C176.939 14.84 177.819 17.656 177.819 21.528C177.819 25.688 176.875 28.744 174.987 30.696C173.099 32.648 170.155 33.624 166.155 33.624ZM166.299 28.008C167.547 28.008 168.363 27.576 168.747 26.712C169.131 25.816 169.323 24.248 169.323 22.008C169.323 20.44 169.243 19.24 169.083 18.408C168.923 17.544 168.619 16.92 168.171 16.536C167.755 16.12 167.099 15.912 166.203 15.912C164.923 15.912 164.187 16.024 163.995 16.248C163.739 16.472 163.531 17.176 163.371 18.36C163.211 19.544 163.131 20.776 163.131 22.056C163.131 24.232 163.339 25.768 163.755 26.664C164.203 27.56 165.051 28.008 166.299 28.008Z'
+                fill='inherit'
+            />
+        </svg>
+    );
+};
