--- src/components/icons/closeIcon/index.ts
@@ -1,3 +1,0 @@
-import CloseIcon from './CloseIcon';
-
-export default CloseIcon;
