--- src/components/icons/gearIcon/index.ts
@@ -1,3 +1,0 @@
-import GearIcon from './GearIcon';
-
-export default GearIcon;
