--- src/components/icons/githubLogo/index.ts
@@ -1,3 +1,0 @@
-import GithubLogo from './GithubLogo';
-
-export default GithubLogo;
