--- src/components/icons/hamburgerIcon/index.ts
@@ -1,3 +1,0 @@
-import HamburgerIcon from './HamburgerIcon';
-
-export default HamburgerIcon;
