+++ src/components/icons/index.ts
@@ -1,0 +1,11 @@
+export * from './CloseIcon';
+export * from './GearIcon';
+export * from './GithubLogo';
+export * from './HamburgerIcon';
+export * from './LinkedinLogo';
+export * from './MoonIcon';
+export * from './PlayIcon';
+export * from './PortfolioLogo';
+export * from './SunIcon';
+export * from './VMark';
+export * from './XMark';
