--- src/components/icons/linkedinLogo/index.ts
@@ -1,3 +1,0 @@
-import LinkedinLogo from './LinkedinLogo';
-
-export default LinkedinLogo;
