--- src/components/icons/moonIcon/index.ts
@@ -1,3 +1,0 @@
-import MoonIcon from './MoonIcon';
-
-export default MoonIcon;
