--- src/components/icons/playIcon/index.ts
@@ -1,3 +1,0 @@
-import PlayIcon from './PlayIcon';
-
-export default PlayIcon;
