--- src/components/icons/portfolioLogo/index.ts
@@ -1,3 +1,0 @@
-import PortfolioLogo from './PortfolioLogo';
-
-export default PortfolioLogo;
