--- src/components/icons/sunIcon/index.ts
@@ -1,3 +1,0 @@
-import SunIcon from './SunIcon';
-
-export default SunIcon;
