--- src/components/icons/vMark/index.ts
@@ -1,3 +1,0 @@
-import VMark from './VMark';
-
-export default VMark;
