--- src/components/icons/xMark/index.ts
@@ -1,3 +1,0 @@
-import XMark from './XMark';
-
-export default XMark;
